Testing with initial props
